backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2,<2.11
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
chromadb>=0.6.3
sentence-transformers>=2.2.2
python-multipart
markitdown[all]>=0.1.0
flask>=2.3.0
requests>=2.31.0
firecrawl==1.7.0
langchain-text-splitters==0.3.8
ollama
# Additional dependencies used in plugins
firecrawl>=1.7.0
pathlib>=1.0.1
uuid>=1.30
markdown2
torch
Pillow
pymupdf
docling
openai